requirements.txt
python-telegram-bot >= 10
